Change type of currency amount struct member

+++ Scripts/GameData/Currency/Currency.cs
@@ -11,6 +11,6 @@ namespace MultiplayerARPG
     public struct CurrencyAmount
     {
         public Currency currency;
-        public float amount;
+        public int amount;
     }
 }
